Added game.py
game.py contains the script to run the game. It leaves methods that are hopefully good enough to train the model with. It contains launch arguments to determine whether it should run headless, run with a player, or run with training mode (just agents or algorithms).
We should be able to use the methods to implement both RL algorithms.