Working server startup, atleast on local
We need to create some frontend that can call the important functions from the python backend. We also need to make some sort of display for the game itself. Python is good enough for simple buttons, but not for the game, so we will probably have to use js or something if we want the game to be better than textual.